requirements.txt
Pillow>=9.5.0
attrdict3>=2.0.2
wxPython>=4.1.0
PyPubSub>=4.0.0
pywin32>=304;platform_system == "Windows"
rarfile>=4.0
pycairo>=1.21.0
configparser>=3.2
natsort>=8
